refactor: publish directly from subproject root, retire stage-npm
Before: the npm package was a transformed copy of the source layout.
`web/movfuscator.mjs` imported `../build/browser/cpp.js`; the demo
imported the wrapper from a sibling; `scripts/stage-npm.sh` rewrote
those imports (`sed s|../build/browser/|./|g`) into `npm-package/` and
flat-copied the lib bundle so npm consumers got the right shape. Two
problems with that:

  1. Source layout != published layout, so what the user installs from
     npm is *not* what lives in the repo. The mapping was a sed away,
     but still extra cognitive load.
  2. `make stage-npm` was the only path to a publishable package. You
     couldn't just `npm publish` from the subproject root.

After: the subproject root *is* the package. Files moved up out of
`web/`:

    web/movfuscator.mjs    → movfuscator.mjs
    web/movfuscator.d.ts   → movfuscator.d.ts
    web/index.html         → index.html
    web/md5.{c,h}          → md5.{c,h}
    web/lib/               → lib/

Wrapper imports go `./build/browser/cpp.js` and `./lib/...` — same paths
resolve in dev (under `make serve`) and from a `node_modules/movfuscator-wasm/`
install. The demo's `import './movfuscator.mjs'` already pointed at a
sibling and continues to work. `stage-deploy.sh` becomes a plain
recursive copy (no sed); `stage-npm.sh` is deleted.

Other tightening:

- `package.json` "files" lists the eight `build/browser/*.{js,wasm}`
  artifacts explicitly + the embedded `package.json` (so emcc's ESM
  loaders parse correctly), instead of `build/browser/` which would
  ship the ~46 intermediate `.o` files from the binutils relink (71
  files → 24 files in the tarball).
- All call sites updated: tests/browser.mjs, tests/run-multi.mjs,
  tests/bench-browser.mjs, scripts/bench.sh, scripts/stage-link-libs.sh,
  Makefile, README, CLAUDE.md.
- `.gitignore`: `web/lib/` → `lib/`; drop `npm-package/`.
- Makefile: drop the `stage-npm` target; `make serve` now opens `/`
  (was `/web/`) and depends on the browser build artifacts.

Verified locally on a clean tree:
  test=19, test-as=19, test-ld=17, test-multi=4, test-browser=55  PASS
  make stage-deploy produces dist/ (27 files / 40 MB)
  npm pack --dry-run: 24 files / 15.9 MB / 42.6 MB unpacked
  tsc 6 strict round-trips the full export surface
