+++ movfuscator-wasm/.gitignore
@@ -2,11 +2,7 @@ vendor/
 build/
 *.i
 
-# Browser ld lazy-loaded payload — generated by scripts/stage-link-libs.sh
+# Lazy-fetched link payload — generated by scripts/stage-link-libs.sh
 # from host /lib32 + /usr/lib32 + /usr/lib/gcc; ~24 MB and host-specific
-# so we don't commit it.
-web/lib/
-
-# npm package staging — generated by scripts/stage-npm.sh, contains the
-# same wasm + lib bundle the deploy ships.
-npm-package/
+# so we don't commit it. Shipped to npm via package.json "files".
+lib/
