+++ movfuscator-wasm/CLAUDE.md
@@ -20,7 +20,7 @@ Two things make this codebase unusual; everything else flows from them:
 
 - **cpp + rcc**: LCC compiled with Emscripten. NODERAWFS for the CLI artifacts in `build/{cpp,rcc}.{js,wasm}` (host pipeline). Browser-mode rebuilds in `build/browser/` use MEMFS + `MODULARIZE` + `EXPORT_ES6`. The browser cpp embeds the small set of `/usr/include` headers our fixtures actually need (collected by [`scripts/collect-headers.sh`](scripts/collect-headers.sh) via `cpp -H`).
 - **as + ld**: GNU binutils 2.44 cross-compiled by [`scripts/build-wasm-as.sh`](scripts/build-wasm-as.sh). Single configure tree produces both. NODERAWFS for the host-pipeline artifacts; the browser variants come from [`scripts/build-wasm-{as,ld}-browser.sh`](scripts/) which re-link the existing .o files with MEMFS flags.
-- **Browser wrapper**: [`web/movfuscator.mjs`](web/movfuscator.mjs) exposes `compile / assemble / link`. The link step lazy-fetches the ~24 MB lib bundle (crt0/crtf/crtd/softfloat32/libc/libm/libgcc/ld-linux) from `web/lib/` on first call. `web/index.html` is the live demo.
+- **Browser wrapper**: [`movfuscator.mjs`](movfuscator.mjs) (at subproject root, with [`movfuscator.d.ts`](movfuscator.d.ts) alongside) exposes `compile / assemble / link / compileToElf / preprocess / compileAsm / parseElfHeader` and `LIB_PATHS`. The link step lazy-fetches the ~24 MB lib bundle (crt0/crtf/crtd/softfloat32/libc/libm/libgcc/ld-linux) from `./lib/` on first call. [`index.html`](index.html) is the live demo, also at subproject root so its `import './movfuscator.mjs'` works both in dev (`make serve`) and on the deployed host. The same subproject root *is* the published npm package — there is no separate staging directory.
 
 ## Things future Claude has bumped into and shouldn't relearn
 
@@ -28,7 +28,7 @@ Two things make this codebase unusual; everything else flows from them:
 - **emcc + binutils**: `bfd/doc/chew` *must* stay a native ELF — emcc otherwise compiles it to a Node shim that fails on `doc.str` mid-build. `build-wasm-as.sh` pre-builds it natively *and* passes `CC_FOR_BUILD=cc` to both configure and make.
 - **`-mx86-used-note=no`** (as) and **`--hash-style=gnu`** (ld) are not optional — they're the deltas between modern binutils defaults and what host `/usr/bin/{as,ld}` emit on Debian 13. Removing them breaks byte-identical tests.
 - **libgcc.a host path varies across distros** — e.g. `/usr/lib/gcc/x86_64-linux-gnu/<ver>/32/libgcc.a`. The wrapper and tests stage it at a *fixed* MEMFS path: `/movfuscator/libgcc.a` (alongside the crt + softfloat objects, all reachable via the default `-L/movfuscator`). [`scripts/stage-link-libs.sh`](scripts/stage-link-libs.sh) resolves the host file via `cc -m32 -print-libgcc-file-name` and copies it to that canonical location.
-- **wasm `ld` browser mode**: avoid `--embed-file` for the ~24 MB link inputs; ship them under `web/lib/` and let the wrapper `fetch()` them lazily (kept in `cachedLibs`). The static path layout under `web/lib/` mirrors the absolute MEMFS paths the link command line uses.
+- **wasm `ld` browser mode**: avoid `--embed-file` for the ~24 MB link inputs; ship them under `lib/` (sibling to the wrapper, gitignored, regenerated by `make stage-link-libs`) and let the wrapper `fetch()` them lazily (kept in `cachedLibs`). The static path layout under `lib/` mirrors the absolute MEMFS paths the link command line uses.
 - **Wrapper hides cpp/rcc/as/ld banners**: their stderr is buffered and only thrown if the tool exits non-zero. Don't reintroduce `console.error` — it lights DevTools red on every successful compile.
 
 ## Further reading
