+++ movfuscator-wasm/tests/bench-browser.mjs
@@ -4,7 +4,7 @@
 // Used by scripts/bench.sh under hyperfine; output is discarded — we only
 // care about how long the full chain takes.
 
-import { compile, assemble, link, LIB_PATHS } from '../web/movfuscator.mjs';
+import { compile, assemble, link, LIB_PATHS } from '../movfuscator.mjs';
 import { readFileSync, readdirSync } from 'node:fs';
 import { dirname, basename, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
@@ -17,7 +17,7 @@ if (process.argv.length < 3) {
 const fixturePath = process.argv[2];
 const fixtureDir = dirname(fixturePath);
 const here = dirname(fileURLToPath(import.meta.url));
-const libDir = join(here, '..', 'web', 'lib');
+const libDir = join(here, '..', 'lib');
 
 const src = readFileSync(fixturePath, 'utf8');
 
@@ -28,7 +28,7 @@ const headers = Object.fromEntries(
         .map(h => [h, readFileSync(join(fixtureDir, h), 'utf8')])
 );
 
-// Link inputs — preloaded from web/lib/ so we don't try to fetch over HTTP.
+// Link inputs — preloaded from ./lib/ so we don't try to fetch over HTTP.
 // LIB_PATHS is re-exported from the wrapper so the bench can't drift from
 // what the production link() actually writes.
 const libs = Object.fromEntries(
